requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic-settings>=2.2
requests>=2.32
httpx>=0.27
pillow>=10.3
python-multipart>=0.0.9
python-dotenv>=1.0
web3>=6.15
py-solc-x==1.1.1
torch>=2.0
torchvision>=0.15
trustmark
slowapi>=0.1.9
ezkl
onnx
ml_dtypes
protobuf
invisible-watermark>=0.2.0
PyWavelets>=1.5
opencv-python-headless>=4.8
arq>=0.25
c2pa-python>=0.5
cryptography>=42.0
transformers>=4.40
